requirements.txt
matplotlib>=3.2.1
numpy>=1.18.5
Pillow>=7.1.2
scipy>=1.4.1